LayoutTests/fast/dom/HTMLLinkElement/prefetch.html

<body>
<script>
if (window.layoutTestController) {
    layoutTestController.waitUntilDone();
    layoutTestController.dumpAsText();
    layoutTestController.dumpResourceResponseMIMETypes();
    setTimeout("layoutTestController.notifyDone()",50);
}
</script>
<html>
<p>This test requires DumpRenderTree to see the log of what resources are loaded.
<link href="prefetch.link" rel="prefetch">
</html>
