LayoutTests/platform/chromium-linux/fast/dom/HTMLLinkElement/prefetch-expected.txt

CONSOLE MESSAGE: line 6: Uncaught TypeError: Object [object Object] has no method 'dumpResourceResponseMIMETypes'

This test requires DumpRenderTree to see the log of what resources are loaded.
